py_proc/requirements.txt
aniso8601
click
Flask
itsdangerous
Jinja2
MarkupSafe
pytz
six
SQLAlchemy
Werkzeug
pandas
numpy
matplotlib.pyplot
seaborn
pandasql